backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.3.2
pymongo==4.8.0
python-telegram-bot==21.5
telethon==1.36.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
aiohttp==3.10.0
sherlock-project==0.15.0
cryptg==0.4.0
